% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{get_gene_bounds}
\alias{get_gene_bounds}
\title{Get gene bounds}
\usage{
get_gene_bounds(gene_name, map_df)
}
\arguments{
\item{gene_name}{name of a gene}

\item{map_df}{data frame mapping gene names to ensembl ID.}
}
\description{
Determines boundaries for plotting a gene by gene name
}
