% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{get_gene_bounds_ensg}
\alias{get_gene_bounds_ensg}
\title{Get gene bounds}
\usage{
get_gene_bounds_ensg(ensg, map_df)
}
\arguments{
\item{ensg}{ensembl ID}

\item{map_df}{data frame mapping gene names to ensembl ID.}
}
\description{
Determines boundaries for plotting a gene  by ensembl ID
}
