% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{make.valid.object}
\alias{make.valid.object}
\title{Builds an object ready to plot}
\usage{
make.valid.object(
  CHR,
  P,
  group = NA,
  BP = NA,
  gene.start = NA,
  gene.end = NA,
  gene.name = NA
)
}
\arguments{
\item{CHR}{numeric column of chromosome numbers}

\item{P}{numeric column of P values to be plotted}

\item{group}{label for plotting, often this will be a tissue name}

\item{BP}{numeric column of base pair locations within a chromosome, either this or gene.start and gene.end are required}

\item{gene.start}{numeric column of base pair locations for the start of a gene}

\item{gene.end}{numeric column of base pair locations for the end of a gene}

\item{gene.name}{name of each gene, this will be shown for P values greater than the tag threshold}
}
\description{
This function take data.frame columns and returns an object ready to plot with nashville.plot()
}
