% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{mirror.plot}
\alias{mirror.plot}
\title{Create a Mirror Plot}
\usage{
mirror.plot(
  gwas = gwas,
  metaxcan = NULL,
  map_df,
  chr = NULL,
  y_min = NULL,
  y_max = NULL,
  y_ticks = NULL,
  zoom_left = 0,
  zoom_right = 0,
  zoom_gene = NULL,
  zoom_ensg = NULL,
  gene_tag_p = NULL,
  color_tissue = NULL,
  sig_line1 = NULL,
  sig_line2 = NULL,
  sig_line1_color = NULL,
  sig_line2_color = NULL,
  draw_bottom = TRUE,
  draw_top = TRUE,
  samp = NULL
)
}
\arguments{
\item{gwas}{data frame of gwas summary stats}

\item{metaxcan}{data frame from read_metaxcan_folder}

\item{chr}{Limit the graph to the given chromosome}

\item{y_min}{numeric, ggplot ymin aes}

\item{y_max}{ggplot ymax aes}

\item{y_ticks}{numeric, number of y breaks}

\item{zoom_left}{plot genes within a gene after this base pair}

\item{zoom_right}{plot genes within a gene before this base pair}

\item{gene_tag_p}{numeric, annotate the SNPs with P-values more extreme than log(gene_tag_p)}

\item{color_tissue}{list of colors to label tissues with}

\item{sig_line1}{numeric, draw a horizontal line at log(sig_line1)}

\item{sig_line2}{numeric, draw a horizontal line at log(sig_line2)}

\item{sig_line1_color}{color for sig_line1}

\item{sig_line2_color}{color for sig_line2}

\item{draw_bottom}{bool whether or not to plot data from gwas}

\item{draw_top}{bool whether or not to plot data from metaxcan}

\item{left}{numeric, leftmost point to plot}

\item{right}{numeric, rightmost point to plot}

\item{test_ylab}{string, y axis label}

\item{x_axis_name}{string, x axis label}

\item{labels_cat}{list, tissue labels}

\item{xbreaks}{numeric, number of x breaks}

\item{xlabels}{list of labels for genes}

\item{draw_genes}{draw lines along the length of each gene}

\item{sample}{draw what percent of points with P-value greater than 0.1}
}
\description{
This function makes two plots of GWAS summary data which share the same
x-axis, one pointed up and the other pointed down.
}
