% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{plot.mh}
\alias{plot.mh}
\title{plot manhattan}
\usage{
\method{plot}{mh}(data, direction, draw_genes)
}
\arguments{
\item{data}{data frame from make.valid.object()}

\item{direction}{direction for data to be drawn}

\item{draw_genes}{bool plot with geam_line}
}
\description{
draws a manhattan plot onto the nashville plot
}
