% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{plot_whole_genome}
\alias{plot_whole_genome}
\title{Generate the nashville plot}
\usage{
plot_whole_genome(
  gwas = gwas,
  metaxcan = metaxcan,
  y_min = NULL,
  y_max = NULL,
  y_ticks = NULL,
  test_ylab = test_ylab,
  x_axis_name,
  gene_tag_p = NULL,
  color_tissue = NULL,
  sig_line1 = NULL,
  sig_line2 = NULL,
  sig_line1_color = NULL,
  sig_line2_color = NULL,
  labels_cat,
  xbreaks,
  xlabels,
  left,
  right,
  draw_genes,
  draw_bottom,
  draw_top
)
}
\arguments{
\item{gwas}{data frame of gwas summary stats}

\item{metaxcan}{data frame from read_metaxcan_folder}

\item{y_min}{numeric, ggplot ymin aes}

\item{y_max}{ggplot ymax aes}

\item{y_ticks}{numeric, number of y breaks}

\item{test_ylab}{string, y axis label}

\item{x_axis_name}{string, x axis label}

\item{gene_tag_p}{numeric, annotate the SNPs with P-values more extreme than log(gene_tag_p)}

\item{color_tissue}{list of colors to label tissues with}

\item{sig_line1}{numeric, draw a horizontal line at log(sig_line1)}

\item{sig_line2}{numeric, draw a horizontal line at log(sig_line2)}

\item{sig_line1_color}{color for sig_line1}

\item{sig_line2_color}{color for sig_line2}

\item{labels_cat}{list, tissue labels}

\item{xbreaks}{numeric, number of x breaks}

\item{xlabels}{list of labels for genes}

\item{left}{numeric, leftmost point to plot}

\item{right}{numeric, rightmost point to plot}

\item{draw_genes}{draw lines along the length of each gene}

\item{draw_bottom}{bool whether or not to plot data from gwas}

\item{draw_top}{bool whether or not to plot data from metaxcan}
}
\description{
This function returns a ggplot2 graph of the nashville plot
}
