% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{read_config}
\alias{read_config}
\title{Read config file}
\usage{
read_config(file)
}
\arguments{
\item{file}{a file path to a 3 column tab separated values file without headers where the first column contains file names, the second column contains group names, and the 3rd column contains colors for those groups to be plotted in.}
}
\description{
Reads in configuration info from tab separated values file
}
