% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{read_metaXcan_folder}
\alias{read_metaXcan_folder}
\title{Read MetaXcan Folder}
\usage{
read_metaXcan_folder(directory, map_df = "37", pattern = "*.csv$")
}
\arguments{
\item{directory}{a string representation of a file path containing predixcan output}

\item{map_df}{a file path containing a mapping gene names to ensembl ID. The header should be "Gene ENSG CHR START_POS END_POS"}

\item{pattern}{a regex describing files to be read from "directory"}

\item{label}{a list of labels to use instead of file names for graphing. These should probably be in alphabetical order.}
}
\description{
This function reads in metaXcan or prediXcan results files by tissue
}
