% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{validate_input}
\alias{validate_input}
\title{Validate Input}
\usage{
validate_input(
  x,
  chr_name = "CHR",
  pos_name = "BP",
  p_val_name = "P",
  snp_name = "SNP",
  ...
)
}
\arguments{
\item{x}{data frame of SNPs}

\item{chr_name}{column name of chromosome number}

\item{pos_name}{column name of SNP position along a chromosome given as number of base pairs}

\item{snp_name}{column name containing SNP names}
}
\description{
This function internally renames the column names for Chromosome, Base Pair
Location, P-value, and SNP name. Cases where Chromosome, Base Pair Location
or P-value are not numeric values are dropped.
}
