% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashvillePlot.R
\name{read_gwas_file}
\alias{read_gwas_file}
\title{Read GWAS File}
\usage{
read_gwas_file(file)
}
\arguments{
\item{file}{path to gwas results as obtained from plink --assoc}
}
\description{
Read GWAS File
}
